move logic to a class library, fix separation
Because of using a separate repository, we need to build a temporary directory
to put all the files required together since nix can't deal with having
both absolute and relative (at least I can't make it work).

This makes it possible for a channel file for example to import other
files using the -p switch.